requirements.txt
streamlit==1.39.0
requests==2.32.3
pandas==2.2.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dateutil==2.9.0.post0
pycountry==24.6.1
